Switch RandomNG class to using std::mt19937_64
Switch the RandomNG to the C++11 standard library random number
implementation. Use the normal, uniform and poisson distributions from the standard library. Add a LandauDistribution distribution compatable with the standard library.

Note that a few places still access the RandGenerator class, so it is
not yet possible to remove the old RNG.