LatticeFunctionTable: Add SetForceLongitudinalStability()
It is often useful to get lattice function for lattices that have no RF
and are therefore unstable in the longitudinal dimension. Merlin
previously offered a bend scale option, however this could be fragile. A
more reliable method is to force R_55 to just below 1. This can be
enabled with:

	LatticeFunctionTable::SetForceLongitudinalStability(true)