Store PSvector::id as an int64_t
ID was previously stored as part of the double[] coordinate array. Now
it is a separate int64_t member.

The ID should be accessed through p.id() not p[6], which will no longer
work.

Note that ID is no longer effected by arithmetic operations between
particles, which is probably what a user would expect. For example

p *= 2;

will double bunch coordinates, but now not the ID.